remove auth and related features
Authentication related feature and routes are removed, specifically:

- auth handlers for email_password and github OAuth providers which were
  working
- other untested OAuth providers (discord, twitter, telegram)
- user-groups, readers, writers: No mixed privacy feature work now

Error messages for the processors like document-{readers, writers} have
been added

http::Request::ud is marked todo!()
