+++ fastn-core/src/http.rs
@@ -1,3 +1,5 @@
+pub const SESSION_COOKIE_NAME: &str = "fastn_session";
+
 #[macro_export]
 macro_rules! server_error {
     ($($t:tt)*) => {{
@@ -273,50 +275,51 @@ impl Request {
             .and_then(|v| v.to_str().map(|v| v.to_string()).ok())
     }
 
-    pub async fn ud(&self, ds: &fastn_ds::DocumentStore) -> Option<fastn_core::UserData> {
-        let session_data = match self.cookie(fastn_core::auth::SESSION_COOKIE_NAME) {
-            Some(c) => {
-                if c.is_empty() {
-                    return None;
-                } else {
-                    c
-                }
-            }
-            None => return None,
-        };
-
-        let session_data = match fastn_core::auth::utils::decrypt(ds, &session_data).await {
-            Ok(v) => v,
-            Err(e) => {
-                tracing::warn!("failed to decrypt session data: {:?}", e);
-                return None;
-            }
-        };
-
-        #[derive(serde::Deserialize)]
-        struct SessionData {
-            session_id: i64,
-            user: fastn_core::UserData,
-        }
-
-        let session_data = match serde_json::from_str::<SessionData>(session_data.as_str()) {
-            Ok(sd) => sd,
-            Err(e) => {
-                tracing::warn!("failed to deserialize session data: {:?}", e);
-                return None;
-            }
-        };
-
-        // if the session does not exist, return None
-        match fastn_core::auth::get_authenticated_user_with_email(&session_data.session_id, ds)
-            .await
-        {
-            Ok(_) => Some(session_data.user),
-            Err(e) => {
-                tracing::warn!("failed to get user data from session: {e}");
-                None
-            }
-        }
+    pub async fn ud(&self, _ds: &fastn_ds::DocumentStore) -> Option<fastn_core::UserData> {
+        todo!("come back to this");
+        // let session_data = match self.cookie(SESSION_COOKIE_NAME) {
+        //     Some(c) => {
+        //         if c.is_empty() {
+        //             return None;
+        //         } else {
+        //             c
+        //         }
+        //     }
+        //     None => return None,
+        // };
+        //
+        // let session_data = match fastn_core::utils::decrypt(ds, &session_data).await {
+        //     Ok(v) => v,
+        //     Err(e) => {
+        //         tracing::warn!("failed to decrypt session data: {:?}", e);
+        //         return None;
+        //     }
+        // };
+        //
+        // #[derive(serde::Deserialize)]
+        // struct SessionData {
+        //     session_id: i64,
+        //     user: fastn_core::UserData,
+        // }
+        //
+        // let session_data = match serde_json::from_str::<SessionData>(session_data.as_str()) {
+        //     Ok(sd) => sd,
+        //     Err(e) => {
+        //         tracing::warn!("failed to deserialize session data: {:?}", e);
+        //         return None;
+        //     }
+        // };
+        //
+        // // if the session does not exist, return None
+        // match fastn_core::auth::get_authenticated_user_with_email(&session_data.session_id, ds)
+        //     .await
+        // {
+        //     Ok(_) => Some(session_data.user),
+        //     Err(e) => {
+        //         tracing::warn!("failed to get user data from session: {e}");
+        //         None
+        //     }
+        // }
     }
 
     pub fn body(&self) -> &[u8] {
