+++ fastn-core/src/library2022/processor/user_group.rs
@@ -1,112 +1,55 @@
 pub fn process(
-    value: ftd_ast::VariableValue,
-    kind: ftd::interpreter::Kind,
-    doc: &ftd::interpreter::TDoc,
-    req_config: &fastn_core::RequestConfig,
+    _value: ftd_ast::VariableValue,
+    _kind: ftd::interpreter::Kind,
+    _doc: &ftd::interpreter::TDoc,
+    _req_config: &fastn_core::RequestConfig,
 ) -> ftd::interpreter::Result<ftd::interpreter::Value> {
-    use itertools::Itertools;
-    let g = req_config
-        .config
-        .package
-        .groups
-        .values()
-        .map(|g| g.to_group_compat())
-        .collect_vec();
-    doc.from_json(&g, &kind, &value)
+    Err(ftd::interpreter::Error::OtherError(
+        "user-groups is not implemented in this version. Switch to an older  \
+            version."
+            .into(),
+    ))
 }
 
 /// processor: user-group-by-id
 pub fn process_by_id(
-    value: ftd_ast::VariableValue,
-    kind: ftd::interpreter::Kind,
-    doc: &ftd::interpreter::TDoc,
-    req_config: &fastn_core::RequestConfig,
+    _value: ftd_ast::VariableValue,
+    _kind: ftd::interpreter::Kind,
+    _doc: &ftd::interpreter::TDoc,
+    _req_config: &fastn_core::RequestConfig,
 ) -> ftd::interpreter::Result<ftd::interpreter::Value> {
-    let headers = match value.get_record(doc.name) {
-        Ok(val) => val.2.to_owned(),
-        Err(e) => return Err(e.into()),
-    };
-
-    let group_id = match headers.get_optional_string_by_key("id", doc.name, value.line_number())? {
-        Some(k) => k,
-        None => {
-            return Err(ftd::interpreter::Error::ParseError {
-                message: "`id` field is mandatory in `user-group-by-id` processor".to_string(),
-                doc_id: doc.name.to_string(),
-                line_number: value.line_number(),
-            })
-        }
-    };
-
-    let g = req_config
-        .config
-        .package
-        .groups
-        .get(group_id.as_str())
-        .map(|g| g.to_group_compat())
-        .ok_or_else(|| ftd::ftd2021::p1::Error::NotFound {
-            key: format!("user-group: `{}` not found", group_id.as_str()),
-            doc_id: doc.name.to_string(),
-            line_number: value.line_number(),
-        })?;
-    doc.from_json(&g, &kind, &value)
+    Err(ftd::interpreter::Error::OtherError(
+        "user-group-by-id is not implemented in this version. Switch to an \
+            older version."
+            .into(),
+    ))
 }
 
 /// processor: get-identities
 /// This is used to get all the identities of the current document
 pub async fn get_identities(
-    value: ftd_ast::VariableValue,
-    kind: ftd::interpreter::Kind,
-    doc: &ftd::interpreter::TDoc<'_>,
-    req_config: &fastn_core::RequestConfig,
+    _value: ftd_ast::VariableValue,
+    _kind: ftd::interpreter::Kind,
+    _doc: &ftd::interpreter::TDoc<'_>,
+    _req_config: &fastn_core::RequestConfig,
 ) -> ftd::interpreter::Result<ftd::interpreter::Value> {
-    use itertools::Itertools;
-
-    let doc_id = fastn_core::library2022::utils::document_full_id(req_config, doc)?;
-
-    let identities =
-        fastn_core::user_group::get_identities(&req_config.config, doc_id.as_str(), true)
-            .await
-            .map_err(|e| ftd::ftd2021::p1::Error::ParseError {
-                message: e.to_string(),
-                doc_id,
-                line_number: value.line_number(),
-            })?;
-
-    Ok(ftd::interpreter::Value::List {
-        data: identities
-            .into_iter()
-            .map(|i| ftd::interpreter::PropertyValue::Value {
-                value: ftd::interpreter::Value::String {
-                    text: i.to_string(),
-                },
-                is_mutable: false,
-                line_number: value.line_number(),
-            })
-            .collect_vec(),
-        kind: ftd::interpreter::KindData {
-            kind,
-            caption: false,
-            body: false,
-        },
-    })
+    Err(ftd::interpreter::Error::OtherError(
+        "get-identities is not implemented in this version. Switch to an \
+            older version."
+            .into(),
+    ))
 }
 
 // is user can_read the document or not based on defined readers in sitemap
 pub async fn is_reader<'a>(
-    value: ftd_ast::VariableValue,
+    _value: ftd_ast::VariableValue,
     _kind: ftd::interpreter::Kind,
-    doc: &ftd::interpreter::TDoc<'a>,
-    req_config: &fastn_core::RequestConfig,
+    _doc: &ftd::interpreter::TDoc<'a>,
+    _req_config: &fastn_core::RequestConfig,
 ) -> ftd::interpreter::Result<ftd::interpreter::Value> {
-    let doc_id = fastn_core::library2022::utils::document_full_id(req_config, doc)?;
-    let is_reader = req_config.can_read(&doc_id, false).await.map_err(|e| {
-        ftd::ftd2021::p1::Error::ParseError {
-            message: e.to_string(),
-            doc_id,
-            line_number: value.line_number(),
-        }
-    })?;
-
-    Ok(ftd::interpreter::Value::Boolean { value: is_reader })
+    Err(ftd::interpreter::Error::OtherError(
+        "is-reader is not implemented in this version. Switch to an older \
+            version."
+            .into(),
+    ))
 }
