+++ fastn-core/src/package/app.rs
@@ -92,60 +92,3 @@ impl AppTemp {
         })
     }
 }
-
-// Takes the path /-/<package-name>/<remaining>/ or /mount-point/<remaining>/
-pub async fn can_read(config: &fastn_core::RequestConfig, path: &str) -> fastn_core::Result<bool> {
-    use itertools::Itertools;
-    // first get the app
-    let readers_groups = if let Some((_, _, _, Some(app))) = config
-        .config
-        .get_mountpoint_sanitized_path(&config.config.package, path)
-    {
-        app.readers.clone()
-    } else {
-        vec![]
-    };
-
-    if readers_groups.is_empty() {
-        return Ok(true);
-    }
-
-    let user_groups = config
-        .config
-        .package
-        .groups
-        .iter()
-        .filter_map(|(id, g)| {
-            if readers_groups.contains(id) {
-                Some(g)
-            } else {
-                None
-            }
-        })
-        .collect_vec();
-
-    let mut app_identities = vec![];
-    for ug in user_groups.iter() {
-        app_identities.extend(ug.get_identities(&config.config).await?)
-    }
-
-    let auth_identities = fastn_core::auth::get_auth_identities(
-        &config.config.ds,
-        config.request.cookies(),
-        app_identities.as_slice(),
-    )
-    .await?;
-
-    fastn_core::user_group::belongs_to(
-        &config.config,
-        user_groups.as_slice(),
-        auth_identities.iter().collect_vec().as_slice(),
-    )
-    .await
-
-    // get the app readers
-    // get the groups
-    // the the access identities
-    // check belongs _to
-    //return Ok(true);
-}
