+++ fastn-core/src/sitemap/mod.rs
@@ -1236,209 +1236,6 @@ impl Sitemap {
         }
     }
 
-    /// This function will return all the readers and readers which are inherited from parent
-
-    // TODO: need to handle special reader: everyone, writer: everyone
-    // readers function should return Vec<UserGroup> or Everyone
-    // it return readers and `confidential`
-    pub fn readers<'a>(
-        &self,
-        doc_path: &str,
-        groups: &'a std::collections::BTreeMap<String, fastn_core::user_group::UserGroup>,
-    ) -> (Vec<&'a fastn_core::user_group::UserGroup>, bool) {
-        use itertools::Itertools;
-
-        for section in self.sections.iter() {
-            let (readers, confidential) = find_section(section, doc_path);
-            if readers.is_empty() {
-                continue;
-            }
-            let readers: Vec<String> = self.readers.iter().cloned().chain(readers).collect();
-            return (
-                readers
-                    .iter()
-                    .unique()
-                    .filter_map(|g| groups.get(g))
-                    .chain(self.writers(doc_path, groups))
-                    .collect(),
-                confidential,
-            );
-        }
-
-        return (vec![], true);
-
-        fn find_toc(toc: &toc::TocItem, doc_path: &str) -> (Vec<String>, bool) {
-            if toc.id.eq(doc_path) {
-                return (toc.readers.clone(), toc.confidential);
-            }
-
-            for child in toc.children.iter() {
-                let (readers, mut confidential) = find_toc(child, doc_path);
-                if readers.is_empty() {
-                    continue;
-                }
-                // Inherit from the parent, if parent is not confidential assuming all the children
-                // are not confidential
-                if !toc.confidential {
-                    confidential = false;
-                }
-
-                return (
-                    readers
-                        .into_iter()
-                        .chain(toc.readers.iter().cloned())
-                        .collect(),
-                    confidential,
-                );
-            }
-            (vec![], true)
-        }
-
-        fn find_subsection(
-            subsection: &section::Subsection,
-            doc_path: &str,
-        ) -> (Vec<String>, bool) {
-            if subsection
-                .id
-                .as_ref()
-                .map(|id| id.eq(doc_path))
-                .unwrap_or(false)
-            {
-                return (subsection.readers.clone(), subsection.confidential);
-            }
-
-            for toc in subsection.toc.iter() {
-                let (readers, mut confidential) = find_toc(toc, doc_path);
-                if readers.is_empty() {
-                    continue;
-                }
-
-                // Inherit from the parent, if parent is not confidential assuming all the children
-                // are not confidential
-                if !subsection.confidential {
-                    confidential = false;
-                }
-
-                return (
-                    readers
-                        .into_iter()
-                        .chain(subsection.readers.iter().cloned())
-                        .collect(),
-                    confidential,
-                );
-            }
-            (vec![], true)
-        }
-
-        fn find_section(section: &section::Section, doc_path: &str) -> (Vec<String>, bool) {
-            if section.id.eq(doc_path) {
-                return (section.readers.clone(), section.confidential);
-            }
-
-            for subsection in section.subsections.iter() {
-                let (readers, mut confidential) = find_subsection(subsection, doc_path);
-                if readers.is_empty() {
-                    continue;
-                }
-                // Inherit from the parent, if parent is not confidential assuming all the children
-                // are not confidential
-                if !section.confidential {
-                    confidential = false;
-                }
-                return (
-                    readers
-                        .into_iter()
-                        .chain(section.readers.iter().cloned())
-                        .collect(),
-                    confidential,
-                );
-            }
-            (vec![], true)
-        }
-    }
-
-    pub fn writers<'a>(
-        &self,
-        doc_path: &str,
-        groups: &'a std::collections::BTreeMap<String, fastn_core::user_group::UserGroup>,
-    ) -> Vec<&'a fastn_core::user_group::UserGroup> {
-        use itertools::Itertools;
-
-        for section in self.sections.iter() {
-            let writers = find_section(section, doc_path);
-            if writers.is_empty() {
-                continue;
-            }
-            let writers: Vec<String> = self.writers.iter().cloned().chain(writers).collect();
-            return writers
-                .iter()
-                .unique()
-                .filter_map(|g| groups.get(g))
-                .collect();
-        }
-
-        return vec![];
-
-        fn find_toc(toc: &toc::TocItem, doc_path: &str) -> Vec<String> {
-            if toc.id.eq(doc_path) {
-                return toc.writers.clone();
-            }
-
-            for child in toc.children.iter() {
-                let writers = find_toc(child, doc_path);
-                if writers.is_empty() {
-                    continue;
-                }
-                return writers
-                    .into_iter()
-                    .chain(toc.writers.iter().cloned())
-                    .collect();
-            }
-            vec![]
-        }
-
-        fn find_subsection(subsection: &section::Subsection, doc_path: &str) -> Vec<String> {
-            if subsection
-                .id
-                .as_ref()
-                .map(|id| id.eq(doc_path))
-                .unwrap_or(false)
-            {
-                return subsection.writers.clone();
-            }
-
-            for toc in subsection.toc.iter() {
-                let writers = find_toc(toc, doc_path);
-                if writers.is_empty() {
-                    continue;
-                }
-                return writers
-                    .into_iter()
-                    .chain(subsection.writers.iter().cloned())
-                    .collect();
-            }
-            vec![]
-        }
-
-        fn find_section(section: &section::Section, doc_path: &str) -> Vec<String> {
-            if section.id.eq(doc_path) {
-                return section.writers.clone();
-            }
-
-            for subsection in section.subsections.iter() {
-                let writers = find_subsection(subsection, doc_path);
-                if writers.is_empty() {
-                    continue;
-                }
-                return writers
-                    .into_iter()
-                    .chain(section.writers.iter().cloned())
-                    .collect();
-            }
-            vec![]
-        }
-    }
-
     /// path: foo/temp/
     /// path: /
     /// This function can be used for if path exists in sitemap or not
