+++ fastn-core/src/utils.rs
@@ -1156,3 +1156,32 @@ async fn get_env_value_or_default(
         }
     }
 }
+
+pub async fn encrypt(ds: &fastn_ds::DocumentStore, input: &str) -> String {
+    use magic_crypt::MagicCryptTrait;
+    let secret_key = secret_key(ds).await;
+    let mc_obj = magic_crypt::new_magic_crypt!(secret_key.as_str(), 256);
+    mc_obj.encrypt_to_base64(input).as_str().to_owned()
+}
+
+pub async fn decrypt(
+    ds: &fastn_ds::DocumentStore,
+    input: &str,
+) -> Result<String, magic_crypt::MagicCryptError> {
+    use magic_crypt::MagicCryptTrait;
+    let secret_key = secret_key(ds).await;
+    let mc_obj = magic_crypt::new_magic_crypt!(&secret_key, 256);
+    mc_obj.decrypt_base64_to_string(input)
+}
+
+pub async fn secret_key(ds: &fastn_ds::DocumentStore) -> String {
+    match ds.env("FASTN_SECRET_KEY").await {
+        Ok(secret) => secret,
+        Err(_e) => {
+            fastn_core::warning!(
+                "WARN: Using default SECRET_KEY. Provide one using FASTN_SECRET_KEY env var."
+            );
+            "FASTN_TEMP_SECRET".to_string()
+        }
+    }
+}
